déplacer la logique de l'import des structures dans un autre fichier
